+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 - Rename `cvd_risk_calculation_sheet_v0` feature flag to `non_lab_based_cvd_risk_calculation_sheet`
 - Add `lab_based_statin_nudge` feature flag.
 - Bug fixes in CI workflows
+- Fetch `lab_based_cvd_risk_calculation_sheet` from remote config.
 
 ### Fixes